+++ mobi.chouette.exchange.netex_stif/src/main/java/mobi/chouette/exchange/netex_stif/parser/DirectionParser.java
@@ -19,7 +19,7 @@ public class DirectionParser implements Parser, Constant {
 		XmlPullParser xpp = (XmlPullParser) context.get(PARSER);
 		NetexStifObjectFactory factory = (NetexStifObjectFactory)context.get(NETEX_STIF_OBJECT_FACTORY);
 		
-		Integer version = (Integer) context.get(VERSION);
+		Long version = (Long) context.get(VERSION);
 		String id = xpp.getAttributeValue(null, ID);
 		Direction direction = factory.getDirection(id);
 		direction.setObjectVersion(version);
